code/describe-and-generate/pyproject.toml
[project]
name = "describe-and-generate"
version = "0.1.0"
description = "Lambda function to analyze images and add AI descriptions to metadata"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.34.0",
    "google-genai>=1.56.0",
    "openai>=1.54.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "imageio-ffmpeg>=0.4.9",
    "supabase>=2.10.0",
]
